Store Networks in Database (+more Networks +fetch from online)

+++ app/src/main/java/org/walleth/data/balances/Balance.kt
@@ -8,7 +8,7 @@ import java.math.BigInteger
 data class Balance(
         val address: Address,
         val tokenAddress: Address,
-        val chain: Long,
+        val chain: BigInteger,
 
         val block: Long,
         val balance: BigInteger
